adding 0 again

+++ 0x11-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int num_nodes = 0;
+	size_t num_nodes = 0;
 
 	if (h == NULL)
 		return (0);
